# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Incident Management System API
  description: Comprehensive API documentation for Ingester, Workflow Engine, Alert Service, and Mock Generator.
  version: 1.0.0

servers:
  - url: http://localhost:8001
    description: Ingester Service
  - url: http://localhost:8002
    description: Workflow Engine
  - url: http://localhost:8004
    description: Alert Service
  - url: http://localhost:8003
    description: Mock Generator

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key

security:
  - ApiKeyAuth: []

paths:
  # --- INGESTER ---
  /api/v1/signals:
    post:
      summary: Ingest a new signal
      tags: [Ingester]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Signal'
      responses:
        '202':
          description: Signal accepted and queued
        '400':
          description: Validation error
        '401':
          description: Unauthorized
        '429':
          description: Backpressure limit reached

  /health:
    get:
      summary: Service health check
      tags: [Shared]
      responses:
        '200':
          description: OK

  # --- WORKFLOW ENGINE ---
  /incidents:
    get:
      summary: List live incidents
      tags: [Workflow Engine]
      responses:
        '200':
          description: Array of work items
  
  /incidents/{id}:
    get:
      summary: Get incident details
      tags: [Workflow Engine]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Incident + RCA + Signal Count

  /incidents/{id}/investigate:
    patch:
      summary: Transition to INVESTIGATING
      tags: [Workflow Engine]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success

  /incidents/{id}/resolve:
    patch:
      summary: Transition to RESOLVED
      tags: [Workflow Engine]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success

  /incidents/{id}/rca:
    post:
      summary: Submit Root Cause Analysis
      tags: [Workflow Engine]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RCA'
      responses:
        '201':
          description: RCA Created

  /incidents/{id}/close:
    patch:
      summary: Transition to CLOSED
      tags: [Workflow Engine]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Success (returns MTTR)
        '422':
          description: RCA Missing or Incomplete

  /incidents/{id}/signals:
    get:
      summary: Get paginated signal history for an incident
      tags: [Workflow Engine]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: List of signals from MongoDB

  /metrics/throughput:
    get:
      summary: Get system throughput metrics
      tags: [Workflow Engine]
      responses:
        '200':
          description: Bucketed metrics data

  # --- ALERT SERVICE ---
  /alerts/send:
    post:
      summary: Manually trigger an alert (Internal)
      tags: [Alert Service]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                incident_id: { type: string }
                component_id: { type: string }
                component_type: { type: string }
                severity: { type: string }
                error_code: { type: string }
                message: { type: string }
                timestamp: { type: string, format: date-time }
      responses:
        '200':
          description: Alert processed

  /alerts/history:
    get:
      summary: Get alert history
      tags: [Alert Service]
      responses:
        '200':
          description: Last 50 alerts

  /alerts/config:
    get:
      summary: Get routing configuration
      tags: [Alert Service]
      responses:
        '200':
          description: Routing rules per component

  /alerts/config/{component_type}:
    put:
      summary: Update routing for a component
      tags: [Alert Service]
      parameters:
        - name: component_type
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                webhook_url: { type: string }
                enabled: { type: boolean }
      responses:
        '200':
          description: Config updated

  # --- MOCK GENERATOR ---
  /simulate/burst:
    post:
      summary: Trigger 200 signal burst
      tags: [Mock Generator]
      responses:
        '200':
          description: Simulation started

  /simulate/outage:
    post:
      summary: Trigger 60s outage simulation
      tags: [Mock Generator]
      responses:
        '200':
          description: Simulation started

schemas:
  Signal:
    type: object
    required: [component_id, component_type, severity, error_code, message, timestamp]
    properties:
      component_id: { type: string }
      component_type: { type: string, enum: [RDBMS, CACHE, API, QUEUE, MCP, NOSQL] }
      severity: { type: string, enum: [P0, P1, P2, P3] }
      error_code: { type: string }
      message: { type: string }
      timestamp: { type: string, format: date-time }
      metadata: { type: object }

  RCA:
    type: object
    required: [root_cause_category, fix_applied, prevention_steps, incident_start, incident_end]
    properties:
      root_cause_category: { type: string, enum: [HARDWARE_FAILURE, SOFTWARE_BUG, CONFIGURATION_ERROR, CAPACITY_EXHAUSTION, NETWORK_ISSUE, HUMAN_ERROR, THIRD_PARTY_FAILURE, UNKNOWN] }
      fix_applied: { type: string, minLength: 10 }
      prevention_steps: { type: string, minLength: 10 }
      incident_start: { type: string, format: date-time }
      incident_end: { type: string, format: date-time }